+++ sdk/eventhub/azure-eventhubs/azure/eventhub/_producer_client.py
@@ -94,6 +94,16 @@ class EventHubProducerClient(EventHubClient):
         :keyword transport_type: The type of transport protocol that will be used for communicating with
          the Event Hubs service. Default is `TransportType.Amqp`.
         :paramtype transport_type: ~azure.eventhub.TransportType
+
+        .. admonition:: Example:
+
+            .. literalinclude:: ../samples/sync_samples/sample_code_eventhub.py
+                :start-after: [START create_eventhub_producer_client_from_conn_str_sync]
+                :end-before: [END create_eventhub_producer_client_from_conn_str_sync]
+                :language: python
+                :dedent: 4
+                :caption: Create a new instance of the EventHubProducerClient from connection string.
+
         """
         return super(EventHubProducerClient, cls).from_connection_string(conn_str, **kwargs)